requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
tensorflow==2.17.1
numpy==1.24.3
Pillow==10.0.0
requests
gunicorn